requirements.txt
python-telegram-bot[ext]>=22.0
requests>=2.25.0
Flask[async]>=2.0.0
nest-asyncio>=1.5.0
pytz
telethon==1.34.0
telethon-secret-chat>=0.1.0
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.9.1
psycopg2-binary>=2.9.0
qrcode[pil]>=7.0.0
Pillow>=9.0.0
Pyrogram>=2.0.0
solana>=0.30.0
solders>=0.18.0
